requirements.txt
lap>=0.5.12,<0.6
numpy>=2.4,<3
onvif-zeep>=0.2.12,<0.3
opencv-python>=4.13,<5
streamlit>=1.56,<2
ultralytics>=8.4,<9
